+++ CHANGELOG.md
@@ -31,6 +31,7 @@ Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangel
 - :tada: **Enhancement** added **`RelativeStandardDeviationCriterion`**
 - :tada: **Enhancement** added **`StandardErrorCriterion`**
 - :tada: **Enhancement** added **`VarianceCriterion`**
+- :tada: **Enhancement** added **`AverageCriterion`**
 
 
 ## 0.14 (released April 25, 2021)